move `Mongofier` to its own file
